refactor : seperate ui & app route - reason : for clean code and server-side logic may needs in app route file
