+++ src/app/authentication/login/page.tsx
@@ -1,16 +1,7 @@
-import AuthenticationForm from '@/components/AuthenticationForm'
+import Login from '@/containers/Authentication/Login'
 
 const LoginPage = () => {
-  return (
-    <div>
-      <AuthenticationForm
-        authenticationType="로그인"
-        authenticationSwitchMessage="아직 회원이 아니라면?"
-        submitButtonMessage="로그인"
-        authenticationSwitchLink="/member/register"
-      />
-    </div>
-  )
+  return <Login />
 }
 
 export default LoginPage
